@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -7,7 +8,7 @@ namespace DotPulsar.Abstractions
     /// <summary>
     /// A consumer abstraction.
     /// </summary>
-    public interface IConsumer : IStateChanged<ConsumerState>, IDisposable
+    public interface IConsumer : IStateChanged<ConsumerState>, IAsyncDisposable
     {
         /// <summary>
         /// Acknowledge the consumption of a single message.
@@ -37,7 +38,7 @@ namespace DotPulsar.Abstractions
         /// <summary>
         /// Receives a single message.
         /// </summary>
-        Task<Message> Receive(CancellationToken cancellationToken = default);
+        IAsyncEnumerable<Message> Receive(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Reset the subscription associated with this consumer to a specific MessageId.
