@@ -7,7 +7,7 @@ namespace DotPulsar.Abstractions
     /// <summary>
     /// A reader abstraction.
     /// </summary>
-    public interface IReader : IStateChanged<ReaderState>, IDisposable
+    public interface IReader : IStateChanged<ReaderState>, IAsyncDisposable
     {
         /// <summary>
         /// Receives a single message.
