@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace DotPulsar.Internal.Abstractions
 {
-    public interface IConsumerStream : IDisposable
+    public interface IConsumerStream : IAsyncDisposable
     {
         Task Send(CommandAck command);
         Task<CommandSuccess> Send(CommandUnsubscribe command);
