@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace DotPulsar.Internal.Abstractions
 {
-    public interface IProducerStream : IDisposable
+    public interface IProducerStream : IAsyncDisposable
     {
         Task<CommandSendReceipt> Send(PulsarApi.MessageMetadata metadata, ReadOnlyMemory<byte> payload);
     }
