@@ -1,4 +1,5 @@
-﻿namespace DotPulsar.Internal.PulsarApi
+﻿#nullable disable
+namespace DotPulsar.Internal.PulsarApi
 {
     // This file was generated by a tool; you should avoid making direct changes.
     // Consider using 'partial classes' to extend these types
@@ -1951,3 +1952,4 @@
 }
 
 #pragma warning restore CS1591, CS0612, CS3021, IDE1006
+#nullable restore
